# docs/swagger.yaml
definitions:
  models.IngredientModel:
    properties:
      condition:
        type: string
      subject:
        type: string
    type: object
  models.RecipeCreateBody:
    properties:
      stages:
        items:
          $ref: '#/definitions/models.StageModel'
        type: array
      title:
        type: string
    type: object
  models.Result:
    properties:
      error_code:
        $ref: '#/definitions/utils.Error'
      error_description:
        type: string
      error_exception:
        type: string
      success:
        type: boolean
    type: object
  models.StageModel:
    properties:
      description:
        type: string
      ingredients:
        items:
          $ref: '#/definitions/models.IngredientModel'
        type: array
      title:
        type: string
    type: object
  models.TokenCheckResult:
    properties:
      expired:
        type: string
      result:
        $ref: '#/definitions/models.Result'
      token:
        type: string
    type: object
  models.UserLoginArgs:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.UserLoginResult:
    properties:
      authentication_token:
        type: string
      id:
        type: string
      result:
        $ref: '#/definitions/models.Result'
      user_infos:
        additionalProperties:
          type: string
        type: object
    type: object
  models.UserRegisterArgs:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      validate_password:
        type: string
    type: object
  models.UserRegisterResult:
    properties:
      result:
        $ref: '#/definitions/models.Result'
    type: object
  utils.Error:
    enum:
    - ERR0303
    - ERR0304
    - ERR0401
    - ERR0402
    - ERR0403
    - ERR0404
    - ERR0405
    - ERR0406
    - ERR0407
    type: string
    x-enum-varnames:
    - ERR0303
    - ERR0304
    - ERR0401
    - ERR0402
    - ERR0403
    - ERR0404
    - ERR0405
    - ERR0406
    - ERR0407
host: 127.0.0.1:8080/
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server on Gorrilla Mux + MongoDB.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MonGorilla Project
  version: "1.0"
paths:
  /api/recipes:
    post:
      consumes:
      - application/json
      description: Creating Recipe in DB with given request body
      operationId: create-new-recipe
      parameters:
      - description: Enter recipe data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RecipeCreateBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: create new recipe
      tags:
      - Users
  /api/users/current_user:
    get:
      consumes:
      - application/json
      description: Token check method for authentication
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenCheckResult'
      security:
      - Bearer: []
      summary: Check validity of token
      tags:
      - Auth
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Login with username and password
      parameters:
      - description: UserLogin
        in: body
        name: userModelArgs
        required: true
        schema:
          $ref: '#/definitions/models.UserLoginArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserLoginResult'
      summary: Login to your account
      tags:
      - Auth
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Register and create account
      parameters:
      - description: UserRegister
        in: body
        name: userModelArgs
        required: true
        schema:
          $ref: '#/definitions/models.UserRegisterArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRegisterResult'
      summary: Create a account
      tags:
      - Auth
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"